pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = "data quality dashboard"
authors = ["Mohamed Abdelaal <mohamed.abdelaal@softwareag.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
numpy = "*"
pandas = "*"
raha = "^1.25"
setuptools = "^69.0.2"
dash = "^2.14.2"
dash-bootstrap-components = "^1.5.0"
pyyaml = "^6.0.1"
jsonlines = "^4.0.0"
impyute = "^0.0.8"
missforest = "^2.2.3"
lightgbm = "^4.1.0"
mlflow = "^2.8.1"
pyodbc = "^5.0.1"
mysql-connector-python = "^8.2.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.2"
deltalake = "^0.13.0"
fastapi = "^0.104.1"
ydata-profiling = "^4.6.2"
torch = {version = "^2.1.1+cpu", source = "pytorch"}
torchvision = {version = "^0.16.1+cpu", source = "pytorch"}
scikit-learn = "~1.3.2"
uvicorn = "^0.24.0.post1"
scikeras = "^0.12.0"
tensorflow = "~2.12"
optuna = "^3.5.0"
levenshtein = "^0.23.0"
gensim = "^4.3.2"
werkzeug = "2.0.3"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
